fix: restructure broadcasting contracts per review
- [contracts] Delete driver.go — consolidate into broadcasting.go
- [contracts] Rename broadcast.go → broadcasting.go
- [contracts] Move Driver interface and ConnectionConfig/PusherOptions/AuthConfig structs into broadcasting.go
- [setup] Simplify env vars to BROADCAST_CONNECTION=log only (PUSHER_* defaults in config stub)
- [tests] Merge 5 separate Channel test functions into TestApplication_Channel with t.Run subtests
- [mocks] Regenerate for updated contracts layout
